requirements.txt
asgiref==3.8.1
boto3==1.35.66
botocore==1.35.66
dj-database-url==0.5.0
Django==5.1.3
django-allauth==65.2.0
django-crispy-forms==2.3
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
s3transfer==0.10.4
sqlparse==0.5.1
stripe==11.2.0
